+++ CHANGELOG.rst
@@ -7,7 +7,9 @@
 
 DEV
 ~~~
+* Cleaned up code to pep8 standards
 * Added recursive list function
+* Added recursive delete function
 * Updated help text in CLI
 
 v1.0.0